pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "static-analyzer"
version = "1.0.0"
description = "Production-quality static analysis framework for Embedded C/C++"
authors = [
    {name = "Static Analyzer Team", email = "dev@staticanalyzer.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["static-analysis", "misra", "cert", "c", "cpp", "embedded", "automotive"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "libclang==16.0.6",
    "pyyaml==6.0.1", 
    "dataclasses-json==0.6.3",
    "typing-extensions==4.9.0",
    "click==8.1.7",
    "jsonschema==4.20.0",
    "openai==1.6.1",
    "colorama==0.4.6"
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-cov==4.1.0",
    "mypy==1.8.0",
    "black==23.12.1",
    "flake8==7.0.0"
]

[project.scripts]
static-analyzer = "static_analyzer.cli:main"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=static_analyzer --cov-report=html --cov-report=term-missing"
